refactor: Consolidate secrets e2e tests into st_app_advanced
Move programmatic secrets tests into the existing st_app_advanced test
instead of having a separate test file. This follows the e2e test best
practice of aggregating related scenario tests to minimize browser loads.
